requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
sqlalchemy==1.4.47
geopy==2.4.1